# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gestión de Inventario y Ventas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /perfumery:
    get:
      operationId: listPerfumery
      tags: [perfumery]
      summary: List all perfumery items
      responses:
        "200":
          description: List of perfumery products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeryItem"
    post:
      operationId: createPerfumeryItem
      tags: [perfumery]
      summary: Create a perfumery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePerfumeryItemBody"
      responses:
        "201":
          description: Created perfumery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerfumeryItem"
        "400":
          description: Bad request
  /perfumery/{id}:
    get:
      operationId: getPerfumeryItem
      tags: [perfumery]
      summary: Get a perfumery item by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Perfumery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerfumeryItem"
        "404":
          description: Not found
    patch:
      operationId: updatePerfumeryItem
      tags: [perfumery]
      summary: Update a perfumery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePerfumeryItemBody"
      responses:
        "200":
          description: Updated perfumery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerfumeryItem"
        "404":
          description: Not found
    delete:
      operationId: deletePerfumeryItem
      tags: [perfumery]
      summary: Delete a perfumery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    PerfumeryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        ml:
          type: integer
        stock:
          type: integer
        costPrice:
          type: number
        salePrice:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - brand
        - ml
        - stock
        - costPrice
        - salePrice
        - createdAt
        - updatedAt
    CreatePerfumeryItemBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        ml:
          type: integer
        stock:
          type: integer
        costPrice:
          type: number
        salePrice:
          type: number
        description:
          type: ["string", "null"]
      required:
        - name
        - brand
        - ml
        - stock
        - costPrice
        - salePrice
    UpdatePerfumeryItemBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        ml:
          type: integer
        stock:
          type: integer
        costPrice:
          type: number
        salePrice:
          type: number
        description:
          type: ["string", "null"]
...